Look for local album art
Sometimes the album art is not available for download or the downloaded image is for a different release (eg cover art from a release in a different country). Now when the user chooses to override the downloaded or missing album art abcde will first look for $ALBUMARTDIR/$ALBUMARTFILE and present that option to the user.